requirements.txt
pillow
fastapi[standard]